+++ agent/src/main/java/com/teamscale/jacoco/agent/options/AgentOptions.java
@@ -3,16 +3,21 @@
 | Copyright (c) 2009-2018 CQSE GmbH                                        |
 |                                                                          |
 +-------------------------------------------------------------------------*/
-package com.teamscale.jacoco.agent;
+package com.teamscale.jacoco.agent.options;
 
 import com.teamscale.client.FileSystemUtils;
 import com.teamscale.client.TeamscaleServer;
+import com.teamscale.jacoco.agent.Agent;
+import com.teamscale.jacoco.agent.AgentBase;
 import com.teamscale.jacoco.agent.commandline.Validator;
+import com.teamscale.jacoco.agent.git_properties.GitPropertiesLocatingTransformer;
+import com.teamscale.jacoco.agent.git_properties.GitPropertiesLocator;
 import com.teamscale.jacoco.agent.store.IUploader;
 import com.teamscale.jacoco.agent.store.NoopUploader;
 import com.teamscale.jacoco.agent.store.UploaderException;
 import com.teamscale.jacoco.agent.store.upload.azure.AzureFileStorageConfig;
 import com.teamscale.jacoco.agent.store.upload.azure.AzureFileStorageUploadStore;
+import com.teamscale.jacoco.agent.store.upload.delay.DelayedCommitDescriptorStore;
 import com.teamscale.jacoco.agent.store.upload.http.HttpUploadStore;
 import com.teamscale.jacoco.agent.store.upload.teamscale.TeamscaleUploader;
 import com.teamscale.jacoco.agent.testimpact.TestExecutionWriter;
@@ -20,14 +25,17 @@ import com.teamscale.jacoco.agent.testimpact.TestwiseCoverageAgent;
 import com.teamscale.jacoco.agent.util.AgentUtils;
 import com.teamscale.jacoco.agent.util.LoggingUtils;
 import com.teamscale.report.EDuplicateClassFileBehavior;
+import com.teamscale.report.testwise.jacoco.cache.CoverageGenerationException;
 import com.teamscale.report.util.ClasspathWildcardIncludeFilter;
 import okhttp3.HttpUrl;
 import org.conqat.lib.commons.assertion.CCSMAssert;
 import org.conqat.lib.commons.collections.PairList;
+import org.jacoco.core.runtime.WildcardMatcher;
 import org.slf4j.Logger;
 
 import java.io.File;
 import java.io.IOException;
+import java.lang.instrument.Instrumentation;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.text.DateFormat;
@@ -36,7 +44,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -66,7 +73,7 @@ public class AgentOptions {
 	/**
 	 * The directory to write the XML traces to.
 	 */
-	/* package */ Path outputDirectory = null;
+	/* package */ Path outputDirectory = AgentUtils.getAgentDirectory().resolve("coverage");
 
 	/**
 	 * The URL to which to upload coverage zips.
@@ -103,12 +110,14 @@ public class AgentOptions {
 	/* package */ boolean shouldIgnoreDuplicateClassFiles = true;
 
 	/**
-	 * Include patterns to pass on to JaCoCo.
+	 * Include patterns for fully qualified class names to pass on to JaCoCo. See {@link WildcardMatcher} for the
+	 * pattern syntax. Individual patterns must be separated by ":".
 	 */
 	/* package */ String jacocoIncludes = null;
 
 	/**
-	 * Exclude patterns to pass on to JaCoCo.
+	 * Exclude patterns for fully qualified class names to pass on to JaCoCo. See {@link WildcardMatcher} for the
+	 * pattern syntax. Individual patterns must be separated by ":".
 	 */
 	/* package */ String jacocoExcludes = null;
 
@@ -133,6 +142,11 @@ public class AgentOptions {
 	/* package */ Integer httpServerPort = null;
 
 	/**
+	 * Whether coverage is written to an exec file in testwise mode or returned via HTTP.
+	 */
+	/* package */ boolean coverageViaHttp = false;
+
+	/**
 	 * The configuration necessary to upload files to an azure file storage
 	 */
 	/* package */ AzureFileStorageConfig azureFileStorageConfig = new AzureFileStorageConfig();
@@ -145,7 +159,7 @@ public class AgentOptions {
 	}
 
 	/**
-	 * Validates the options and throws an exception if they're not valid.
+	 * Validates the options and returns a validator with all validation errors.
 	 */
 	/* package */ Validator getValidator() {
 		Validator validator = new Validator();
@@ -154,10 +168,7 @@ public class AgentOptions {
 			validator.isTrue(path.canRead(), "Path '" + path + "' is not readable");
 		}
 
-		validator.ensure(() -> {
-			CCSMAssert.isNotNull(outputDirectory, "You must specify an output directory");
-			FileSystemUtils.ensureDirectoryExists(outputDirectory.toFile());
-		});
+		validator.ensure(() -> FileSystemUtils.ensureDirectoryExists(outputDirectory.toFile()));
 
 		if (loggingConfig != null) {
 			validator.ensure(() -> {
@@ -172,9 +183,6 @@ public class AgentOptions {
 			});
 		}
 
-		validator.isTrue(!useTestwiseCoverageMode() || uploadUrl == null, "'upload-url' option is " +
-				"incompatible with Testwise coverage mode!");
-
 		validator.isFalse(uploadUrl == null && !additionalMetaDataFiles.isEmpty(),
 				"You specified additional meta data files to be uploaded but did not configure an upload URL");
 
@@ -183,7 +191,7 @@ public class AgentOptions {
 
 		validator.isTrue((azureFileStorageConfig.hasAllRequiredFieldsSet() || azureFileStorageConfig
 						.hasAllRequiredFieldsNull()),
-				"If you want to upload data to an azure file storage you need to provide both " +
+				"If you want to upload data to an Azure file storage you need to provide both " +
 						"'azure-url' and 'azure-key' ");
 
 		List<Boolean> configuredStores = Stream
@@ -191,9 +199,37 @@ public class AgentOptions {
 						uploadUrl != null).filter(x -> x).collect(Collectors.toList());
 
 		validator.isTrue(configuredStores.size() <= 1, "You cannot configure multiple upload stores, " +
-				"such as a teamscale instance, upload url or azure file storage");
+				"such as a Teamscale instance, upload URL or Azure file storage");
+
+		appendTestwiseCoverageValidations(validator);
 
 		return validator;
+	}
+
+	private void appendTestwiseCoverageValidations(Validator validator) {
+		validator.isFalse(useTestwiseCoverageMode() && httpServerPort == null && testEnvironmentVariable == null,
+				"You use 'mode' 'TESTWISE' but did neither use 'http-server-port' nor 'test-env'! One of them is required!");
+
+		validator.isFalse(useTestwiseCoverageMode() && httpServerPort != null && testEnvironmentVariable != null,
+				"You did set both 'http-server-port' and 'test-env'! Only one of them is allowed!");
+
+		validator.isFalse(useTestwiseCoverageMode() && uploadUrl != null, "'upload-url' option is " +
+				"incompatible with Testwise coverage mode!");
+
+		validator.isFalse(useTestwiseCoverageMode() && !teamscaleServer.hasAllRequiredFieldsNull(),
+				"'teamscale-' options are incompatible with Testwise coverage mode!");
+
+		validator.isFalse(useTestwiseCoverageMode() && coverageViaHttp && classDirectoriesOrZips.isEmpty(),
+				"You use 'coverage-via-http' but did not provide any class files via 'class-dir'!");
+
+		validator.isFalse(!useTestwiseCoverageMode() && httpServerPort != null,
+				"You use 'http-server-port' but did not set 'mode' to 'TESTWISE'!");
+
+		validator.isFalse(!useTestwiseCoverageMode() && coverageViaHttp,
+				"You use 'coverage-via-http' but did not set 'mode' to 'TESTWISE'!");
+
+		validator.isFalse(!useTestwiseCoverageMode() && testEnvironmentVariable != null,
+				"You use 'test-env' but did not set 'mode' to 'TESTWISE'!");
 	}
 
 	/**
@@ -208,7 +244,9 @@ public class AgentOptions {
 			builder.append(",excludes=").append(jacocoExcludes);
 		}
 
-		if (classDirectoriesOrZips.isEmpty()) {
+		// Don't dump class files in testwise mode when coverage is written to an exec file
+		boolean needsClassFiles = mode == EMode.NORMAL || coverageViaHttp;
+		if (classDirectoriesOrZips.isEmpty() && needsClassFiles) {
 			Path tempDir = createTemporaryDumpDirectory();
 			tempDir.toFile().deleteOnExit();
 			builder.append(",classdumpdir=").append(tempDir.toAbsolutePath().toString());
@@ -247,7 +285,7 @@ public class AgentOptions {
 
 	/** Sets output to none for normal mode and destination file in testwise coverage mode */
 	private String getModeSpecificOptions() {
-		if (useTestwiseCoverageMode()) {
+		if (useTestwiseCoverageMode() && !coverageViaHttp) {
 			return "sessionid=,destfile=" + getTempFile("jacoco", "exec").getAbsolutePath();
 		} else {
 			return "output=none";
@@ -264,22 +302,28 @@ public class AgentOptions {
 	 * Returns in instance of the agent that was configured. Either an agent with interval based line-coverage dump or
 	 * the HTTP server is used.
 	 */
-	public AgentBase createAgent() throws UploaderException {
+	public AgentBase createAgent(
+			Instrumentation instrumentation) throws UploaderException, CoverageGenerationException {
 		if (useTestwiseCoverageMode()) {
 			return new TestwiseCoverageAgent(this, new TestExecutionWriter(getTempFile("test-execution", "json")));
 		} else {
-			return new Agent(this);
+			return new Agent(this, instrumentation);
 		}
 	}
 
 	/**
 	 * Creates an uplaoder for the coverage XMLs.
 	 */
-	public IUploader createUploader() throws UploaderException {
+	public IUploader createUploader(Instrumentation instrumentation) throws UploaderException {
 		if (uploadUrl != null) {
 			return new HttpUploadStore(uploadUrl, additionalMetaDataFiles);
 		}
 		if (teamscaleServer.hasAllRequiredFieldsSet()) {
+			if (teamscaleServer.commit == null) {
+				logger.info("You did not provide a commit to upload to directly, so the Agent will try and" +
+						" auto-detect it by searching all profiled Jar/War/Ear/... files for a git.properties file.");
+				return createDelayedTeamscaleStore(instrumentation);
+			}
 			return new TeamscaleUploader(teamscaleServer);
 		}
 
@@ -289,6 +333,16 @@ public class AgentOptions {
 		}
 
 		return new NoopUploader();
+	}
+
+	private IUploader createDelayedTeamscaleStore(Instrumentation instrumentation)
+			throws UploaderException {
+
+		DelayedCommitDescriptorStore store = new DelayedCommitDescriptorStore(
+				commit -> new TeamscaleUploader(teamscaleServer), outputDirectory);
+		GitPropertiesLocator locator = new GitPropertiesLocator(store);
+		instrumentation.addTransformer(new GitPropertiesLocatingTransformer(locator, getLocationIncludeFilter()));
+		return store;
 	}
 
 	/**
@@ -304,6 +358,13 @@ public class AgentOptions {
 	}
 
 	/**
+	 * Get the directory to which the coverage files are written to
+	 */
+	public Path getOutputDirectory() {
+		return outputDirectory;
+	}
+
+	/**
 	 * @see #dumpIntervalInMinutes
 	 */
 	public int getDumpIntervalInMinutes() {
@@ -313,7 +374,7 @@ public class AgentOptions {
 	/**
 	 * @see #shouldIgnoreDuplicateClassFiles
 	 */
-	public EDuplicateClassFileBehavior duplicateClassFileBehavior() {
+	public EDuplicateClassFileBehavior getDuplicateClassFileBehavior() {
 		if (shouldIgnoreDuplicateClassFiles) {
 			return EDuplicateClassFileBehavior.WARN;
 		} else {
@@ -358,7 +419,7 @@ public class AgentOptions {
 	 * @see #jacocoIncludes
 	 * @see #jacocoExcludes
 	 */
-	public Predicate<String> getLocationIncludeFilter() {
+	public ClasspathWildcardIncludeFilter getLocationIncludeFilter() {
 		return new ClasspathWildcardIncludeFilter(jacocoIncludes, jacocoExcludes);
 	}
 
@@ -370,6 +431,11 @@ public class AgentOptions {
 	/** Whether coverage should be dumped on JVM shutdown. */
 	public boolean shouldDumpOnExit() {
 		return shouldDumpOnExit;
+	}
+
+	/** Whether coverage should be dumped via http. */
+	public boolean shouldDumpCoverageViaHttp() {
+		return coverageViaHttp;
 	}
 
 	/** Describes the two possible modes the agent can be started in. */