+++ agent/src/main/java/com/teamscale/jacoco/agent/store/upload/teamscale/TeamscaleUploader.java
@@ -23,19 +23,9 @@ public class TeamscaleUploader implements IUploader {
 	/** Teamscale server details. */
 	private final TeamscaleServer teamscaleServer;
 
-	/** The API which performs the upload. */
-	private final ITeamscaleService api;
-
 	/** Constructor. */
 	public TeamscaleUploader(TeamscaleServer teamscaleServer) {
 		this.teamscaleServer = teamscaleServer;
-
-		api = TeamscaleServiceGenerator.createService(
-				ITeamscaleService.class,
-				teamscaleServer.url,
-				teamscaleServer.userName,
-				teamscaleServer.userAccessToken
-		);
 	}
 
 	@Override
@@ -54,6 +44,13 @@ public class TeamscaleUploader implements IUploader {
 		logger.debug("Uploading JaCoCo artifact to {}", teamscaleServer);
 
 		try {
+			// Cannot be executed in the constructor as this causes issues in WildFly server (See #100)
+			ITeamscaleService api = TeamscaleServiceGenerator.createService(
+					ITeamscaleService.class,
+					teamscaleServer.url,
+					teamscaleServer.userName,
+					teamscaleServer.userAccessToken
+			);
 			api.uploadReport(
 					teamscaleServer.project,
 					teamscaleServer.commit,