+++ agent/src/test/java/com/teamscale/jacoco/agent/convert/ConverterTest.java
@@ -1,8 +1,9 @@
 package com.teamscale.jacoco.agent.convert;
 
 import org.conqat.lib.commons.filesystem.FileSystemUtils;
-import org.conqat.lib.commons.test.CCSMTestCaseBase;
-import org.junit.Test;
+import org.conqat.lib.commons.test.ManagedTestData;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
 
 import java.io.File;
 import java.util.Collections;
@@ -10,17 +11,18 @@ import java.util.Collections;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /** Basic smoke test for the converter. */
-public class ConverterTest extends CCSMTestCaseBase {
+public class ConverterTest {
+
+	private ManagedTestData managedTestData = new ManagedTestData(getClass());
 
 	/**
-	 * Ensures that running the converter on valid input does not yield any errors
-	 * and produces a coverage XML report.
+	 * Ensures that running the converter on valid input does not yield any errors and produces a coverage XML report.
 	 */
 	@Test
-	public void testSmokeTest() throws Exception {
-		File execFile = useTestFile("coverage.exec");
-		File classFile = useTestFile("TestClass.class");
-		File outputFile = createTmpFile("coverage.xml", "");
+	public void testSmokeTest(@TempDir File tempDir) throws Exception {
+		File execFile = managedTestData.useDataFile("coverage.exec");
+		File classFile = managedTestData.useDataFile("TestClass.class");
+		File outputFile = new File(tempDir, "coverage.xml");
 
 		ConvertCommand arguments = new ConvertCommand();
 		arguments.inputFiles = Collections.singletonList(execFile.getAbsolutePath());